refactor and change service implementation

+++ internal/goal/dto/create/create_goal_request.go
@@ -4,6 +4,6 @@ type CreateGoalRequest struct {
 	Title         string               `json:"title" validate:"required,max=255"`
 	Description   string               `json:"description,omitempty"`
 	HoursPerWeek  int                  `json:"hours_per_week" validate:"required,min=1"`
-	EstimatedTime int64                `json:"estimated_time"`
+	EstimatedTime int                  `json:"estimated_time"`
 	Phases        []CreatePhaseRequest `json:"phases,omitempty"`
 }
